Perfect Proj.#2(Total 100)

+++ src/userprog/exception.c
@@ -149,7 +149,9 @@ page_fault (struct intr_frame *f)
   not_present = (f->error_code & PF_P) == 0;
   write = (f->error_code & PF_W) != 0;
   user = (f->error_code & PF_U) != 0;
-  if (fault_addr <= (uintptr_t)4 || fault_addr >= (uintptr_t)PHYS_BASE){
+
+  // if this thread accesse the abnormal stack pointer, thread_exit call!
+  if ((int)fault_addr <= 4 || (int)fault_addr >= (int)PHYS_BASE){
     struct thread *t = thread_current();
     struct member *member = lookup_child(t->tid);
     if (member) {
@@ -157,6 +159,7 @@ page_fault (struct intr_frame *f)
       member->exit_status = -1;
       sema_up(&member->sema);
     }
+
     printf("%s: exit(%d)\n", t->file_name, -1);
     thread_exit();
     return;
